etl/requirements.txt
pandas
requests
sqlalchemy
pyodbc
python-dotenv
